+++ src/components/index.tsx
@@ -1,0 +1,8 @@
+import ContestList from "<@>/pages/contests";
+import React from "react";
+
+const Contests: React.FC = () => {
+  return <ContestList />;
+};
+
+export default Contests;
